tenant server tests

+++ packages/server/routes/tenant.test.js
@@ -14,29 +14,37 @@ const {
 afterAll(db.close);
 
 beforeAll(async () => {
-  await db.query(`drop schema if exists test2 cascade`);
-  await db.query(`drop schema if exists peashoot cascade`);
+  if (!db.isSQLite) {
+    await db.query(`drop schema if exists test2 cascade`);
+    await db.query(`drop schema if exists peashoot cascade`);
+  }
 });
 
 describe("tenant routes", () => {
-  it("creates tenant", async () => {
-    db.enable_multi_tenant();
-    const app = await getApp({ disableCsrf: true });
-    await request(app)
-      .post("/tenant/create")
-      .send("subdomain=test2")
-      .expect(toInclude("Success"));
-    expect(1).toBe(1);
-  });
-  it("creates tenant with capital letter", async () => {
-    db.enable_multi_tenant();
-    const app = await getApp({ disableCsrf: true });
-    await request(app)
-      .post("/tenant/create")
-      .send("subdomain=Peashoot")
-      .expect(toInclude("Success"));
-    db.set_sql_logging(false);
+  if (!db.isSQLite) {
+    it("creates tenant", async () => {
+      db.enable_multi_tenant();
+      const app = await getApp({ disableCsrf: true });
+      await request(app)
+        .post("/tenant/create")
+        .send("subdomain=test2")
+        .expect(toInclude("Success"));
+      expect(1).toBe(1);
+    });
+    it("creates tenant with capital letter", async () => {
+      db.enable_multi_tenant();
+      const app = await getApp({ disableCsrf: true });
+      await request(app)
+        .post("/tenant/create")
+        .send("subdomain=Peashoot")
+        .expect(toInclude("Success"));
+      db.set_sql_logging(false);
 
-    expect(1).toBe(1);
-  });
+      expect(1).toBe(1);
+    });
+  } else {
+    it("does not support tenants on SQLite", async () => {
+      expect(db.isSQLite).toBe(true);
+    });
+  }
 });
